range-finder-api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytest
inference-sdk
